pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-incident_response_triage_env"
version = "0.1.0"
description = "Incident Response Triage Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    # Environment-specific dependencies
    # Add all dependencies needed for your environment here
    # Examples:
    # "numpy>=1.19.0",
    # "torch>=2.0.0",
    # "gymnasium>=0.29.0",
    # "openspiel>=1.0.0",
    # "smolagents>=1.22.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m incident_response_triage_env.server.app
server = "incident_response_triage_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = [
    "incident_response_triage_env",
    "incident_response_triage_env.server",
    "incident_response_triage_env.graders",
    "incident_response_triage_env.scenarios",
]
package-dir = { "incident_response_triage_env" = ".", "incident_response_triage_env.server" = "server", "incident_response_triage_env.graders" = "graders", "incident_response_triage_env.scenarios" = "scenarios" }